Refactor many definitions out of topics and into folders within mobject
