+++ mobject/geometry.py
@@ -1,11 +1,13 @@
+from __future__ import absolute_import
+
 from constants import *
 
 import itertools as it
 import numpy as np
 
 from mobject.mobject import Mobject
-from mobject.vectorized_mobject import VGroup
-from mobject.vectorized_mobject import VMobject
+from mobject.types.vectorized_mobject import VGroup
+from mobject.types.vectorized_mobject import VMobject
 from utils.bezier import interpolate
 from utils.config_ops import digest_config
 from utils.config_ops import digest_locals
@@ -495,18 +497,20 @@ class Arrow(Line):
             stroke_color = self.color,
             stroke_width = 0,
         )
+        tip.add_at_end = add_at_end
         self.set_tip_points(tip, add_at_end, preserve_normal = False)
         self.add(tip)
         if not hasattr(self, 'tip'):
-            self.tip = []
-        self.tip.append(tuple((tip, add_at_end)))
+            self.tip = VGroup()
+            self.tip.match_style(tip)
+        self.tip.add(tip)
         return tip
 
     def add_rectangular_stem(self):
         self.rect = Rectangle(
             stroke_width = 0,
-            fill_color = self.tip[0][0].get_fill_color(),
-            fill_opacity = self.tip[0][0].get_fill_opacity()
+            fill_color = self.tip.get_fill_color(),
+            fill_opacity = self.tip.get_fill_opacity()
         )
         self.add_to_back(self.rect)
         self.set_stroke(width = 0)
@@ -515,7 +519,7 @@ class Arrow(Line):
     def set_rectangular_stem_points(self):
         start, end = self.get_start_and_end()
         vect = end - start
-        tip_base_points = self.tip[0][0].get_anchors()[1:]
+        tip_base_points = self.tip[0].get_anchors()[1:]
         tip_base = center_of_mass(tip_base_points)
         tbp1, tbp2 = tip_base_points
         perp_vect = tbp2 - tbp1
@@ -576,7 +580,7 @@ class Arrow(Line):
         return self
 
     def get_normal_vector(self):
-        p0, p1, p2 = self.tip[0][0].get_anchors()
+        p0, p1, p2 = self.tip[0].get_anchors()
         result = np.cross(p2 - p1, p1 - p0)
         norm = np.linalg.norm(result)
         if norm == 0:
@@ -590,7 +594,7 @@ class Arrow(Line):
 
     def get_end(self):
         if hasattr(self, "tip"):
-            return self.tip[0][0].get_anchors()[0]
+            return self.tip[0].get_anchors()[0]
         else:
             return Line.get_end(self)
 
@@ -599,7 +603,7 @@ class Arrow(Line):
 
     def put_start_and_end_on(self, *args, **kwargs):
         Line.put_start_and_end_on(self, *args, **kwargs)
-        self.set_tip_points(self.tip[0][0], preserve_normal = False)
+        self.set_tip_points(self.tip[0], preserve_normal = False)
         self.set_rectangular_stem_points()
         return self
 
@@ -607,8 +611,7 @@ class Arrow(Line):
         Line.scale(self, scale_factor, **kwargs)
         if self.preserve_tip_size_when_scaling:
             for t in self.tip:
-                print(t)
-                self.set_tip_points(t[0], add_at_end=t[1])
+                self.set_tip_points(t, add_at_end=t.add_at_end)
         if self.use_rectangular_stem:
             self.set_rectangular_stem_points()
         return self
@@ -691,88 +694,6 @@ class Square(Rectangle):
             width = self.side_length,
             **kwargs
         )
-
-class SurroundingRectangle(Rectangle):
-    CONFIG = {
-        "color" : YELLOW,
-        "buff" : SMALL_BUFF,
-    }
-    def __init__(self, mobject, **kwargs):
-        digest_config(self, kwargs)
-        kwargs["width"] = mobject.get_width() + 2*self.buff
-        kwargs["height"] = mobject.get_height() + 2*self.buff
-        Rectangle.__init__(self, **kwargs)
-        self.move_to(mobject)
-
-class BackgroundRectangle(SurroundingRectangle):
-    CONFIG = {
-        "color" : BLACK,
-        "stroke_width" : 0,
-        "fill_opacity" : 0.75,
-        "buff" : 0
-    }
-    def __init__(self, mobject, **kwargs):
-        SurroundingRectangle.__init__(self, mobject, **kwargs)
-        self.original_fill_opacity = self.fill_opacity
-
-    def pointwise_become_partial(self, mobject, a, b):
-        self.set_fill(opacity = b*self.original_fill_opacity)
-        return self
-
-    def get_fill_color(self):
-        return Color(self.color)
-
-class ScreenRectangle(Rectangle):
-    CONFIG = {
-        "width_to_height_ratio" : 16.0/9.0,
-        "height" : 4,
-    }
-    def generate_points(self):
-        self.width = self.width_to_height_ratio * self.height
-        Rectangle.generate_points(self)
-
-class FullScreenRectangle(ScreenRectangle):
-    CONFIG = {
-        "height" : FRAME_HEIGHT,
-    }
-
-class FullScreenFadeRectangle(FullScreenRectangle):
-    CONFIG = {
-        "stroke_width" : 0,
-        "fill_color" : BLACK,
-        "fill_opacity" : 0.7,
-    }
-
-class PictureInPictureFrame(Rectangle):
-    CONFIG = {
-        "height" : 3,
-        "aspect_ratio" : (16, 9)
-    }
-    def __init__(self, **kwargs):
-        digest_config(self, kwargs)
-        height = self.height
-        if "height" in kwargs:
-            kwargs.pop("height")
-        Rectangle.__init__(
-            self,
-            width = self.aspect_ratio[0],
-            height = self.aspect_ratio[1],
-            **kwargs
-        )
-        self.scale_to_fit_height(height)
-
-class Cross(VGroup):
-    CONFIG = {
-        "stroke_color" : RED,
-        "stroke_width" : 6,
-    }
-    def __init__(self, mobject, **kwargs):
-        VGroup.__init__(self, 
-            Line(UP+LEFT, DOWN+RIGHT),
-            Line(UP+RIGHT, DOWN+LEFT),
-        )
-        self.replace(mobject, stretch = True)
-        self.set_stroke(self.stroke_color, self.stroke_width)
 
 class Grid(VMobject):
     CONFIG = {
