+++ mobject/svg/tex_mobject.py
@@ -2,11 +2,11 @@ from constants import *
 
 from svg_mobject import SVGMobject
 from svg_mobject import VMobjectFromSVGPathstring
-from topics.geometry import BackgroundRectangle
+from mobject.shape_matchers import BackgroundRectangle
 from utils.config_ops import digest_config
-from vectorized_mobject import VGroup
-from vectorized_mobject import VMobject
-from vectorized_mobject import VectorizedPoint
+from mobject.types.vectorized_mobject import VGroup
+from mobject.types.vectorized_mobject import VMobject
+from mobject.types.vectorized_mobject import VectorizedPoint
 
 import collections
 import operator as op
@@ -230,68 +230,6 @@ class TextMobject(TexMobject):
         "template_tex_file" : TEMPLATE_TEXT_FILE,
         "alignment" : "\\centering",
     }
-
-class Brace(TexMobject):
-    CONFIG = {
-        "buff" : 0.2,
-        "width_multiplier" : 2,
-        "max_num_quads" : 15,
-        "min_num_quads" : 0,
-    }
-    def __init__(self, mobject, direction = DOWN, **kwargs):
-        digest_config(self, kwargs, locals())
-        angle = -np.arctan2(*direction[:2]) + np.pi
-        mobject.rotate(-angle, about_point = ORIGIN)
-        left  = mobject.get_corner(DOWN+LEFT)
-        right = mobject.get_corner(DOWN+RIGHT)
-        target_width = right[0]-left[0]
-
-        ## Adding int(target_width) qquads gives approximately the right width
-        num_quads = np.clip(
-            int(self.width_multiplier*target_width),
-            self.min_num_quads, self.max_num_quads
-        )
-        tex_string = "\\underbrace{%s}"%(num_quads*"\\qquad")
-        TexMobject.__init__(self, tex_string, **kwargs)
-        self.tip_point_index = np.argmin(self.get_all_points()[:,1])
-        self.stretch_to_fit_width(target_width)
-        self.shift(left - self.get_corner(UP+LEFT) + self.buff*DOWN)
-        for mob in mobject, self:
-            mob.rotate(angle, about_point = ORIGIN)
-
-    def put_at_tip(self, mob, use_next_to = True, **kwargs):
-        if use_next_to:
-            mob.next_to(
-                self.get_tip(),
-                np.round(self.get_direction()),
-                **kwargs
-            )
-        else:
-            mob.move_to(self.get_tip())
-            buff = kwargs.get("buff", DEFAULT_MOBJECT_TO_MOBJECT_BUFFER)
-            shift_distance = mob.get_width()/2.0+buff
-            mob.shift(self.get_direction()*shift_distance)
-        return self
-
-    def get_text(self, *text, **kwargs):
-        text_mob = TextMobject(*text)
-        self.put_at_tip(text_mob, **kwargs)
-        return text_mob
-
-    def get_tex(self, *tex, **kwargs):
-        tex_mob = TexMobject(*tex)
-        self.put_at_tip(tex_mob, **kwargs)
-        return tex_mob
-
-    def get_tip(self):
-        # Very specific to the LaTeX representation
-        # of a brace, but it's the only way I can think
-        # of to get the tip regardless of orientation.
-        return self.get_all_points()[self.tip_point_index]
-
-    def get_direction(self):
-        vect = self.get_tip() - self.get_center()
-        return vect/np.linalg.norm(vect)
 
 class BulletedList(TextMobject):
     CONFIG = {
