+++ mobject/types/image_mobject.py
@@ -10,7 +10,6 @@ from random import random
 from constants import *
 
 from mobject.mobject import Mobject
-from mobject.point_cloud_mobject import PMobject
 from utils.bezier import interpolate
 from utils.color import color_to_int_rgb
 from utils.color import interpolate_color
