+++ once_useful_constructs/fractals.py
@@ -3,15 +3,15 @@ from animation.transform import Transform
 from pi_creature.pi_creature import PiCreature
 from pi_creature.pi_creature import Randolph
 from pi_creature.pi_creature import get_all_pi_creature_modes
-from mobject.vectorized_mobject import VGroup
-from mobject.vectorized_mobject import VMobject
-from mobject.vectorized_mobject import VectorizedPoint
+from mobject.types.vectorized_mobject import VGroup
+from mobject.types.vectorized_mobject import VMobject
+from mobject.types.vectorized_mobject import VectorizedPoint
 from scene.scene import Scene
-from topics.geometry import Circle
-from topics.geometry import Line
-from topics.geometry import Polygon
-from topics.geometry import RegularPolygon
-from topics.geometry import Square
+from mobject.geometry import Circle
+from mobject.geometry import Line
+from mobject.geometry import Polygon
+from mobject.geometry import RegularPolygon
+from mobject.geometry import Square
 from utils.bezier import interpolate
 from utils.color import color_gradient
 from utils.config_ops import digest_config
