+++ topics/graph_scene.py
@@ -1,19 +1,19 @@
 from constants import *
 
 from scene.scene import Scene
-# from topics.geometry import 
+# from mobject.geometry import 
 from animation.creation import ShowCreation
 from animation.update import UpdateFromAlphaFunc
 from animation.creation import Write
 from animation.transform import Transform
-from mobject.tex_mobject import TexMobject
-from mobject.tex_mobject import TextMobject
-from mobject.vectorized_mobject import VGroup
-from mobject.vectorized_mobject import VectorizedPoint
+from mobject.svg.tex_mobject import TexMobject
+from mobject.svg.tex_mobject import TextMobject
+from mobject.types.vectorized_mobject import VGroup
+from mobject.types.vectorized_mobject import VectorizedPoint
 from topics.functions import ParametricFunction
-from topics.geometry import DashedLine
-from topics.geometry import Line
-from topics.geometry import Rectangle
+from mobject.geometry import DashedLine
+from mobject.geometry import Line
+from mobject.geometry import Rectangle
 from topics.number_line import NumberLine
 from utils.bezier import interpolate
 from utils.color import color_gradient
