+++ topics/vector_space_scene.py
@@ -9,18 +9,18 @@ from animation.transform import ApplyPointwiseFunction
 from animation.creation import FadeOut
 from animation.transform import Transform
 from mobject.mobject import Mobject
-from mobject.tex_mobject import TexMobject
-from mobject.tex_mobject import TextMobject
-from mobject.vectorized_mobject import VGroup
-from mobject.vectorized_mobject import VMobject
+from mobject.svg.tex_mobject import TexMobject
+from mobject.svg.tex_mobject import TextMobject
+from mobject.types.vectorized_mobject import VGroup
+from mobject.types.vectorized_mobject import VMobject
 from scene.scene import Scene
-from topics.geometry import Arrow
-from topics.geometry import BackgroundRectangle
-from topics.geometry import Circle
-from topics.geometry import Dot
-from topics.geometry import Line
-from topics.geometry import Square
-from topics.geometry import Vector
+from mobject.geometry import Arrow
+from mobject.shape_matchers import BackgroundRectangle
+from mobject.geometry import Circle
+from mobject.geometry import Dot
+from mobject.geometry import Line
+from mobject.geometry import Square
+from mobject.geometry import Vector
 from topics.number_line import Axes
 from topics.number_line import NumberPlane
